+++ packages/interactions/src/types/Providers/AWSLexV2Provider.ts
@@ -10,6 +10,16 @@
  * CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
  * and limitations under the License.
  */
-export * from './AWSLexProvider';
-export * from './AWSLexV2Provider';
-export * from './InteractionsProvider';
+export interface AWSLexV2ProviderOption {
+	name: string;
+	botId: string;
+	aliasId: string;
+	localeId: string;
+	region: string;
+	providerName: string;
+	onComplete?(botname: string, callback: (err, confirmation) => void): void;
+}
+
+export interface AWSLexV2ProviderOptions {
+	[key: string]: AWSLexV2ProviderOption;
+}